+++ MapboxVisionTests/Mocks/MockFrameRecorder.swift
@@ -2,6 +2,10 @@ import Foundation
 @testable import MapboxVision
 
 class MockFrameRecorder: NSObject, FrameRecorder {
+    func stopRecording() {
+        actionLog.append(.stopRecording)
+    }
+
     enum Action: Equatable {
         case startRecording
         case stopRecording
@@ -12,10 +16,6 @@ class MockFrameRecorder: NSObject, FrameRecorder {
 
     func startRecording(to path: String, settings: VideoSettings) {
         actionLog.append(.startRecording)
-    }
-
-    func stopRecording(completion: (() -> Void)?) {
-        actionLog.append(.stopRecording)
     }
 
     func handle(frame: CMSampleBuffer) {