+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Vision
 - Removed location update in background
 - Fixed a crash on receiving location updates while stopping `VisionManager`
+- Fixed a crash due to race condition in `ObservableVideoSource`
 
 ### AR